Adding a script which runs multiple tests.
This script generates a number of random keys of 128, 192 and 256 bit
lengths (all 3 AES flavors) and files of various sizes with random
data.  It then encrypts and decrypts all files with all keys and
compares the input plain text with the output plain text.  Failures
occur when the input plain text and the output plain text do not
match.

This should probably be broken up in to multiple test cases.